feat(config): add a `noSpinner` option to disable the spinner when running commands
The issue with spinners is that they refresh the terminal for each
frame, causes problems for scripts that require input from the user as
the prompt for the input is erased.

The `noSpinner` option allows the user to disable a spinner for a
specific script, which will prevent the screen from being refreshed.